refactor: 상수를 커스텀 훅별로 분리
