+++ hooks/src/lib/index.ts
@@ -1,5 +1,4 @@
 export { default as useCardNumbers } from './cardNumber/useCardNumbers';
 export { default as useExpiryDate } from './expiryDate/useExpiryDate';
 export { default as useCVC } from './cvc/useCVC';
-export * from './constants';
 export * from './types';
